Chaotix HUD Added!
Chaotix HUD for Sonic the Hedgehog (2013) is a mod made by me, which ported the assets from Sonic 3: Angel Island Revisited to Sonic the Hedgehog (2013), and no, the mod is not a sprite replacement. It adds the sprites separately (it makes use of scripts) and for the Special Stage Results, it has a separate “.gif” file (which also makes use of scripts). That said, I hope you enjoy the mod.